+++ CHANGES.md
@@ -46,6 +46,8 @@
 -   core dom: Add ``wrap`` wrap an element with a wrapper element.
 -   core dom: Add ``hide`` and ``show`` for DOM elements which retain the original display value.
 -   core dom: Add ``find_parents`` to find all parents of an element matching a CSS selector.
+-   core dom: Add ``find_scoped`` to search for elements matching the given selector within the current scope of the given element
+    unless an ``id`` selector is given - in that case the search is done globally.
 -   pat date picker: Support updating a date if it is before another dependent date.
 -   pat tabs: Refactor based on ``ResizeObserver`` and fix problems calculating the with with transitions.
 -   pat tabs: When clicking on the ``extra-tabs`` element, toggle between ``open`` and ``closed`` classes to allow opening/closing an extra-tabs menu via CSS.